Issue #69: Change `range` to require only one arg
Currently range is hard-coded to take both max and min. This is
unhelpful in cases where we may only want to check one of the bounds.

Update so that it behaves more like the `length` validator and can take
either max or min, as well as both. Also ensure that at least one of
them is supplied.