Update error handling to use proc-macro-error
This significantly improves the error experience. Instead of the
proc-macro panicking and returning an error message to that effect, this
change presents the error as a compiler error.

Further changes can be made to provide help comments to the proc-macro
error messages to further improve the user experience.

The proc-macro-error library is a wrapper around the unstable
Diagnostics API on nightly and the stable compile_error! macro on
stable. So, if a consumer of validator_derive is using nightly, they'll
get slightly better error messages than a user on stable. Once the
Diagnostics API is stabilized (if it ever does), the library will
seamlessly use it, requiring little, if any, effort from validator
maintainers.