Add `sensitive` meta parameter for validators
Often, struct fields contain sensitive data which should never be
displayed once the struct has been initially constructed or user input
has been parsed (e.g. passwords and credit card numbers). This is
especially problematic when logging validation errors as the invalid
values may be potentially leaked in plaintext to logs.

This change adds a boolean `sensitive` validator "meta parameter" similar
to the `message` and `code` parameters. This marks a field as sensitive
such that the value should not be included in validation errors.