Drop old not_a_string_type test
This test was checking that the derive macro assertion about the type
not being one of the known string types was working, but now since we
allow implementations of the email validation for third-party types
there is no way for us to perform this check in the derive macro.

The same assertion is still used for the credit card and non-control
character validators, but those will also soon be switched to be
trait-based.